尝试引入LRU 已删除快照，改为当进行fdb时，对数据加读锁，并复制map，释放读锁后，立即对拷贝进行写文件操作。 新增并发测试结果记录 忽略压测文件，避免操作系统不同导致的问题，之后统一采用makefile中的命令生成

+++ cache/state.go
@@ -19,6 +19,6 @@ func (s *State) Del(k string, v []byte) {
 }
 
 //计算出占用内存
-func (s *State) Memory() uint64 {
-	return uint64(8 * (s.KeySize + s.ValueSize))
+func (s *State) Memory() int64 {
+	return 8 * (s.KeySize + s.ValueSize)
 }
